Fix comparison editor
