+++ State Machines/Parameters/Parameters/ObjectParameter.cs
@@ -7,5 +7,9 @@ namespace Shears.StateMachines
         public ObjectParameter(string name) : base(name)
         {
         }
+
+        public ObjectParameter(string name, Object value) : base(name, value)
+        {
+        }
     }
 }
