+++ State Machines/Shears.StateMachines.asmdef
@@ -2,7 +2,8 @@
     "name": "Shears.StateMachines",
     "rootNamespace": "Shears.StateMachines",
     "references": [
-        "GUID:b28a31c08beef024d83a7aa1ec3f2f6d"
+        "GUID:b28a31c08beef024d83a7aa1ec3f2f6d",
+        "GUID:f9343658ba9f34f40af4f1a7b0c76be7"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
