+++ State Machines/StateMachineBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0e0298050bed4cf4f9c1ff67265fd331
